pyproject.toml
[project]
name = "nonebot-plugin-alchelper"
version = "0.2.0"
description = "以nonebot-plugin-alconna为基础提供系列帮助功能的插件"
authors = [
    {name = "RF-Tar-Railt",email = "3165388245@qq.com"},
]
dependencies = [
    "nonebot2>=2.0.1",
    "nonebot-plugin-alconna>=0.26.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
[tool.pdm.build]
includes = ["src"]
